Replace the naive recursion with Longstaff-Schwartz
The per-path recursion from the original paper decides exercise using each
path's own realised future. That is perfect foresight, and it priced the
benchmark call at 14.96 against an exact value of 10.4506 -- a 43% error
that grew without bound in m (+56% by m=21).

LSM replaces the decision with a cross-sectional estimate: at each
exercise date, regress the realised discounted future cashflow on a
quadratic basis in moneyness over the in-the-money paths, and exercise
where intrinsic beats the fitted continuation. The regression supplies
the decision only -- valuation always uses the realised cashflow, since
substituting the fit back in reintroduces bias. Node m needs no
regression: holding there is worth the European value over the remaining
dt, an exact conditional expectation.

Result at N=1e6: all five backends price the benchmark call at 10.4502
(LCG) and 10.4506 (QMC) against the exact 10.4506.

Validation. A call on a non-dividend stock cannot distinguish "LSM works"
from "LSM never exercises", because never exercising is the right answer.
So this adds put support (OptionParams::type, bs_put) and an independent
CRR binomial tree, restricted to the same m exercise dates so both price
the identical Bermudan rather than comparing a Bermudan to an American.
LSM lands within 0.14% of the tree across four strike/vol cases, with a
+1.22 early-exercise premium that would be zero if exercise were broken.

CPU/GPU comparability. The device LCG, Moro inverse CND and Sobol were
single precision while the host was double, so the two sides computed
different functions and no test could tell a GPU bug from expected drift.
All three are now double, and the GPU and host QMC backends agree to
1e-6. On a consumer GPU with 1/64-rate FP64 this costs throughput; a
comparison that means something is worth more.

Structure. Option math shared by host and device moved to
core/hd_math.hpp. The backward pass is core/lsm.cpp on the CPU (one
implementation for serial and OpenMP, differing only by -fopenmp) and
cuda/lsm_gpu.cu on the GPU. Path storage is now required -- N x (m+1)
doubles, ~176 MB at N=1e6 -- because the regression is cross-sectional.
The device regression reduces per block and sums on the host in block
order rather than using atomicAdd on doubles, which would make the price
depend on block completion order.

Also: all CUDA calls now go through CUDA_TRY error macros, and the
serial-vs-OpenMP check was replaced with a one-thread-vs-many check,
which is what that binary can actually verify.

+++ src/cuda/moro_device.cuh
@@ -1,36 +1,44 @@
 #pragma once
 #include <cuda_runtime.h>
 
-__device__ __forceinline__ float moro_inv_cnd_device(float u) {
-    const float a0 =  2.50662823884f;
-    const float a1 = -18.61500062529f;
-    const float a2 =  41.39119773534f;
-    const float a3 = -25.44106049637f;
-    const float b0 = -8.47351093090f;
-    const float b1 =  23.08336743743f;
-    const float b2 = -21.06224101826f;
-    const float b3 =   3.13082909833f;
-    const float c0 = 0.3374754822726147f;
-    const float c1 = 0.9761690190917186f;
-    const float c2 = 0.1607979714918209f;
-    const float c3 = 0.0276438810333863f;
-    const float c4 = 0.0038405729373609f;
-    const float c5 = 0.0003951896511349f;
-    const float c6 = 0.0000321767881768f;
-    const float c7 = 0.0000002888167364f;
-    const float c8 = 0.0000003960315187f;
+// Moro's inverse cumulative normal, in double precision.
+//
+// This deliberately mirrors src/core/moro_inv_cnd.cpp coefficient for
+// coefficient. The earlier single-precision version carried around 1e-7 of tail
+// error -- larger than the QMC convergence advantage it was meant to help
+// measure -- and made the CPU and GPU backends incomparable, so no test could
+// tell a real GPU bug from expected precision drift. The cost is a slower
+// kernel; a trustworthy comparison is the point of the project.
+__device__ __forceinline__ double moro_inv_cnd_device(double u) {
+    const double a0 =   2.50662823884;
+    const double a1 = -18.61500062529;
+    const double a2 =  41.39119773534;
+    const double a3 = -25.44106049637;
+    const double b0 =  -8.47351093090;
+    const double b1 =  23.08336743743;
+    const double b2 = -21.06224101826;
+    const double b3 =   3.13082909833;
+    const double c0 = 0.3374754822726147;
+    const double c1 = 0.9761690190917186;
+    const double c2 = 0.1607979714918209;
+    const double c3 = 0.0276438810333863;
+    const double c4 = 0.0038405729373609;
+    const double c5 = 0.0003951896511349;
+    const double c6 = 0.0000321767881768;
+    const double c7 = 0.0000002888167364;
+    const double c8 = 0.0000003960315187;
 
-    float x = u - 0.5f;
-    float r;
-    if (fabsf(x) < 0.42f) {
+    double x = u - 0.5;
+    double r;
+    if (fabs(x) < 0.42) {
         r = x * x;
-        r = x * (((a3*r + a2)*r + a1)*r + a0) /
-               ((((b3*r + b2)*r + b1)*r + b0)*r + 1.0f);
+        r = x * (((a3 * r + a2) * r + a1) * r + a0) /
+               ((((b3 * r + b2) * r + b1) * r + b0) * r + 1.0);
     } else {
-        r = (x > 0.0f) ? logf(-logf(1.0f - u)) : logf(-logf(u));
-        r = c0 + r*(c1 + r*(c2 + r*(c3 + r*(c4 +
-            r*(c5 + r*(c6 + r*(c7 + r*c8)))))));
-        if (x < 0.0f) r = -r;
+        r = (x > 0.0) ? log(-log(1.0 - u)) : log(-log(u));
+        r = c0 + r * (c1 + r * (c2 + r * (c3 + r * (c4 +
+            r * (c5 + r * (c6 + r * (c7 + r * c8)))))));
+        if (x < 0.0) r = -r;
     }
     return r;
 }
